pyproject.toml
[project]
name = "starterkit"
version = "0.1.0"
requires-python = ">=3.11,<3.12"
dependencies = [
  "pydantic>=2.7,<3.0",
  "python-dotenv>=1.0,<2.0",
  "structlog>=24,<25",
  "langgraph>=0.2,<0.4",
]

[project.optional-dependencies]
test = [
  "pytest>=8,<9",
  "pytest-cov>=4.0,<5.0",
  "pytest-mock>=3.10,<4.0"
]
agent = [
  "langchain>=0.3,<0.4",
  "langchain-openai>=0.2,<0.4",
  "langchain-community>=0.2,<0.4",
]
dev = [
  "pre-commit>=3.0,<4.0",
  "ruff>=0.3.0,<1.0",
  "safety>=2.0,<3.0"
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
